requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn
cv2
h5py
tqdm
scipy
scipy.io
PIL.Image
os
glob
matplotlib
keras
tensorflow
json
torch


# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs

# utilities
six>=1.14
joblib
memoized-property
termcolor

#CSRNet libraries
h5py
scipy
os
glob
json
matplotlib
image
CSRNet
torch
tqdm
